backend/requirements.txt
# Core Framework
fastapi[standard]==0.121.2

# Database
sqlalchemy==2.0.44
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Validation & Settings
pydantic==2.12.4
pydantic-settings>=2.1.0
PyJWT==2.10.1
